Put index wrapper struct in header
