+++ jni/src/nmslib_wrapper.cpp
@@ -25,17 +25,6 @@
 #include <string>
 
 
-struct IndexWrapper {
-    explicit IndexWrapper(const string& spaceType) {
-        // Index gets constructed with a reference to data (see above) but is otherwise unused
-        similarity::ObjectVector data;
-        space.reset(similarity::SpaceFactoryRegistry<float>::Instance().CreateSpace(spaceType, similarity::AnyParams()));
-        index.reset(similarity::MethodFactoryRegistry<float>::Instance().CreateMethod(false, "hnsw", spaceType, *space, data));
-    }
-    std::unique_ptr<similarity::Space<float>> space;
-    std::unique_ptr<similarity::Index<float>> index;
-};
-
 std::string TranslateSpaceType(const std::string& spaceType);
 
 void knn_jni::nmslib_wrapper::CreateIndex(knn_jni::JNIUtilInterface * jniUtil, JNIEnv * env, jintArray idsJ,
@@ -160,9 +149,9 @@ jlong knn_jni::nmslib_wrapper::LoadIndex(knn_jni::JNIUtilInterface * jniUtil, JN
     }
 
     // Load index
-    IndexWrapper * indexWrapper;
+    knn_jni::nmslib_wrapper::IndexWrapper * indexWrapper;
     try {
-        indexWrapper = new IndexWrapper(spaceTypeCpp);
+        indexWrapper = new knn_jni::nmslib_wrapper::IndexWrapper(spaceTypeCpp);
         indexWrapper->index->LoadIndex(indexPathCpp);
         indexWrapper->index->SetQueryTimeParams(similarity::AnyParams(queryParams));
     } catch (...) {
@@ -184,7 +173,7 @@ jobjectArray knn_jni::nmslib_wrapper::QueryIndex(knn_jni::JNIUtilInterface * jni
         throw std::runtime_error("Invalid pointer to index");
     }
 
-    auto *indexWrapper = reinterpret_cast<IndexWrapper*>(indexPointerJ);
+    auto *indexWrapper = reinterpret_cast<knn_jni::nmslib_wrapper::IndexWrapper*>(indexPointerJ);
 
     int dim	= jniUtil->GetJavaFloatArrayLength(env, queryVectorJ);
 
@@ -223,7 +212,7 @@ jobjectArray knn_jni::nmslib_wrapper::QueryIndex(knn_jni::JNIUtilInterface * jni
 }
 
 void knn_jni::nmslib_wrapper::Free(jlong indexPointerJ) {
-    auto *indexWrapper = reinterpret_cast<IndexWrapper*>(indexPointerJ);
+    auto *indexWrapper = reinterpret_cast<knn_jni::nmslib_wrapper::IndexWrapper*>(indexPointerJ);
     delete indexWrapper;
 }
